lb int

+++ visual.py
@@ -250,8 +250,8 @@ def print_stock(stock_name):
 
 def print_leaderboard():
     df = pd.read_csv(f"{SEASON_ID}/all_investors.csv", index_col='name')
-    df['Cash balance ($)'] = df.apply(lambda x:round(get_balance(x.name)), axis=1)
-    df['Net worth ($)'] = df.apply(lambda x:round(get_net_worth(x.name)), axis=1)
+    df['Cash balance ($)'] = df.apply(lambda x:int(round(get_balance(x.name))), axis=1)
+    df['Net worth ($)'] = df.apply(lambda x:int(round(get_net_worth(x.name))), axis=1)
     df = df.sort_values(by='Net worth ($)', ascending=False)
     df.insert(0,'Name', df.index)
     return df[['Name','Net worth ($)','Cash balance ($)']]
